some project structural changes, correctected registration form import
