+++ app/forms/login_form.py
@@ -3,6 +3,7 @@ from wtforms import StringField, PasswordField, SubmitField, BooleanField
 from wtforms.validators import DataRequired, Email, Length
 from wtforms.widgets import PasswordInput
 
+
 class LoginForm(FlaskForm):
     email = StringField(
         'Email Address',
