+++ app/models/password_validator.py
@@ -1,6 +1,7 @@
 from string import punctuation as valid_symbols
 from wtforms import ValidationError
 
+
 class CustomPasswordValidator():
     def __init__(self):
         self.message = 'Password must have at least one upper letter, one lower letter, one symbol, one number and no space.'
