+++ app/routes/routes.py
@@ -1,49 +1,9 @@
-from flask import render_template, redirect, url_for, request
+from flask import render_template, redirect, url_for
 from flask_login import login_user, login_required, logout_user, current_user
-from app.models.user import User
-from app.models.car import Car
-from db import db
-from app.forms.login_form import LoginForm
-from app.forms.car_form import CarForm
-# from app.forms.test_form import TestForm
+from app import db, User, Car, LoginForm, CarForm
+from werkzeug.security import check_password_hash
+from app import select_where
 
-from werkzeug.security import generate_password_hash, check_password_hash
-
-
-from sqlalchemy.sql.elements import BinaryExpression
-from sqlalchemy import select
-
-
-# TODO: move to utils or smth
-class SelectWrapper:
-    def __init__(self, model_class, *expressions):
-        self.model_class = model_class
-        self.expressions = expressions
-
-    def statement(self):
-        return select(self.model_class).where(*self.expressions)
-
-    def one_or_none(self):
-        return db.session.execute(self.statement()).scalar_one_or_none()
-
-    def all(self):
-        return db.session.execute(self.statement()).scalars().all()
-
-    def first(self):
-        return db.session.execute(self.statement()).scalars().first()
-
-    def count(self):
-        return len(self.all())
-
-
-
-
-def select_where(*expressions : BinaryExpression):
-    column = expressions[0].left
-    model = column._annotations['parententity'] # voodoo
-    return SelectWrapper(model, *expressions)
-
-# app/routes/routes.py
 
 # TODO: Create blueprints, etc.
 def register_routes(app):
@@ -52,7 +12,7 @@ def register_routes(app):
     def login():
         form = LoginForm()
         if form.validate_on_submit():
-            user = select_where(User.email == form.email.data).one_or_none()
+            user : User = select_where(User.email == form.email.data).one_or_none()
             if user and check_password_hash(user.password_hash, form.password.data):
                 login_user(user)
                 return redirect(url_for('car_list'))
@@ -67,7 +27,7 @@ def register_routes(app):
     @app.route('/cars')
     @login_required
     def car_list():
-        cars = Car.query.filter_by(user_id=current_user.id).all()
+        cars : list[Car] = Car.query.filter_by(user_id=current_user.id).all()
         return render_template('car_list.html', cars=cars)
 
     @app.route('/cars/add', methods=['GET', 'POST'])
